FacturacionElectronica/requirements.txt
# Librerias necesarias para el funcionamiento del programa, con el siguiente comando en la terminal se instalan todas las librerias necesarias
# pip install -r requirements.txt

openpyxl
pandas
python-dotenv